break robot_display into two renderables
